riscv support.

+++ ld-elf2flt.c
@@ -24,7 +24,6 @@
 #include <unistd.h>
 #include <errno.h>
 #include <libiberty.h>
-#include <filenames.h>
 
 #include "stubs.h"
 const char *elf2flt_progname;
@@ -568,7 +567,8 @@ int main(int argc, char *argv[])
 	if (!flag_verbose) {
 		unlink(tmp_file);
 		unlink(output_flt);
-		unlink(output_elf);
+		if (output_elf)
+			unlink(output_elf);
 	} else {
 		fprintf(stderr,
 			"leaving elf2flt temp files behind:\n"
